# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "T&C API",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/clinics": {
            "get": {
                "tags": [
                    "Clinics"
                ],
                "summary": "Get All Clinics",
                "description": "Displays a listing of all clinics.",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All clinics fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All clinics fetched successfully!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClinicResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "pagination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clinics"
                ],
                "summary": "Create a Clinic",
                "description": "Stores a newly created clinic.",
                "operationId": "store",
                "requestBody": {
                    "description": "Clinic details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "country_id",
                                    "category_ids"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Clinic Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Clinic Address"
                                    },
                                    "postcode": {
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Clinic City"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "Clinic Latitude"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "Clinic Longitude"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clinic created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clinic created successfully!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ClinicResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryResource": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "en": {
                        "type": "string"
                    },
                    "hr": {
                        "type": "string"
                    },
                    "subcategories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        }
                    }
                },
                "type": "object"
            },
            "ClinicResource": {
                "title": "Clinic Resource",
                "description": "Represents a clinic in the system.",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the clinic.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the clinic.",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the clinic.",
                        "type": "string"
                    },
                    "address": {
                        "description": "The address of the clinic.",
                        "type": "string"
                    },
                    "postcode": {
                        "description": "The postcode of the clinic.",
                        "type": "string"
                    },
                    "city": {
                        "description": "The city where the clinic is located.",
                        "type": "string"
                    },
                    "country_name": {
                        "description": "The name of the country where the clinic is located.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "The phone number of the clinic.",
                        "type": "string"
                    },
                    "mobile": {
                        "description": "The mobile phone number of the clinic.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the clinic.",
                        "type": "string",
                        "format": "email"
                    },
                    "web": {
                        "description": "The website of the clinic.",
                        "type": "string"
                    },
                    "contact_person": {
                        "description": "The contact person of the clinic.",
                        "type": "string"
                    },
                    "contact_phone": {
                        "description": "The contact phone number of the clinic.",
                        "type": "string"
                    },
                    "contact_email": {
                        "description": "The contact email address of the clinic.",
                        "type": "string",
                        "format": "email"
                    },
                    "latitude": {
                        "description": "The latitude of the clinic location.",
                        "type": "string"
                    },
                    "longitude": {
                        "description": "The longitude of the clinic location.",
                        "type": "string"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "updated_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "logo_image": {
                        "description": "The path to the logo image of the clinic.",
                        "type": "string"
                    },
                    "categories": {
                        "description": "The categories associated with the clinic.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User Resource",
                "description": "Represents a user.",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the user.",
                        "type": "string",
                        "format": "email"
                    },
                    "clinic": {
                        "$ref": "#/components/schemas/ClinicResource"
                    },
                    "role": {
                        "description": "The role of the user.",
                        "type": "string"
                    },
                    "permissions": {
                        "description": "The permissions assigned to the user.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}